packages/telemetry/pyproject.toml
[project]
name = "snip-telemetry"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
    "opentelemetry-semantic-conventions>=0.41b0",
    "structlog>=24.0",
]

[project.optional-dependencies]
fastapi = ["opentelemetry-instrumentation-fastapi>=0.41b0"]
sqlalchemy = ["opentelemetry-instrumentation-sqlalchemy>=0.41b0"]
httpx = ["opentelemetry-instrumentation-httpx>=0.41b0"]
grpc = ["opentelemetry-instrumentation-grpc>=0.41b0"]
dev = [
    "snip-telemetry[fastapi,sqlalchemy,httpx,grpc]",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "pyright>=1.1.400",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "UP006",
    "UP007",
    "UP035",
    "UP045",
    "UP047",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
exclude = ["tests"]
reportMissingImports = "warning"
